Add split text method
